feat(470): improved response structure BREAKING CHANGE: response structure is different
The validator does well with invalidating a poorly structure
configuration. However, it was unclear how the caller could respond
to an error returned from the validator. Was the field supposed to be
a string? Was the field required? Was the input even valid in the first
place?

This change addresses it by distinguishing what is an error case and
what is a valid parse attempt. In addition, it enumerates the errors
encountered while validating the template configuration.